components/openai_client.py
import os
import openai
from dotenv import load_dotenv
from typing import Dict, Any

# Load environment variables from .env file
load_dotenv()

class OpenAIClient:
    def __init__(self):
        # Retrieve API key from environment variable
        self.api_key = os.getenv("OPENAI_API_KEY")
        if not self.api_key:
            raise ValueError("OPENAI_API_KEY not found in environment variables. Please set it in your .env file.")
        
        openai.api_key = self.api_key

    def get_realtime_news(self, query: str, num_results: int = 5) -> Dict[str, Any]:
        """Fetches real-time news using OpenAI (simulated for now, would integrate with a news API)"""
        try:
            # In a real application, this would involve:
            # 1. Calling a real-time news API (e.g., NewsAPI, Google News API) to get raw news articles.
            # 2. Using OpenAI's GPT models to summarize or extract key information from these articles.
            
            # For demonstration, we'll simulate a response, but the structure shows how GPT could be used.
            
            # Example of how you might use OpenAI to process news (conceptual):
            # response = openai.chat.completions.create(
            #     model="gpt-3.5-turbo",
            #     messages=[
            #         {"role": "system", "content": "You are a helpful assistant that summarizes news articles."},
            #         {"role": "user", "content": f"Summarize recent news about {query}."}
            #     ]
            # )
            # return {"success": True, "news_summary": response.choices[0].message.content}

            # Placeholder for real-time news for Iran
            if "Iran" in query:
                return {
                    "success": True,
                    "query": query,
                    "results": [
                        {
                            "title": "Iran's Latest Political Developments",
                            "source": "Reuters",
                            "date": "2025-07-03",
                            "summary": "Recent reports indicate ongoing diplomatic discussions and internal policy shifts in Iran."
                        },
                        {
                            "title": "Economic Sanctions and Their Impact on Iran",
                            "source": "Bloomberg",
                            "date": "2025-07-02",
                            "summary": "Analysis of the current economic situation in Iran under international sanctions."
                        }
                    ]
                }
            else:
                return {
                    "success": True,
                    "query": query,
                    "results": [
                        {
                            "title": f"Recent developments in {query}",
                            "source": "Simulated News Agency",
                            "date": "2025-07-03",
                            "summary": f"This is a simulated news summary for {query}. Real-time integration would fetch actual data."
                        }
                    ]
                }
        except Exception as e:
            return {"success": False, "error": str(e)}

    def get_travel_advisory(self, country: str) -> Dict[str, Any]:
        """Fetches real-time travel advisory using OpenAI (simulated for now, would integrate with a travel API)"""
        try:
            # Similar to news, this would involve calling a real travel advisory API
            # and potentially using OpenAI to process or reformat the information.

            if country == "Iran":
                return {
                    "success": True,
                    "country": country,
                    "advisory": "High degree of caution due to regional tensions and internal security concerns. Review your travel insurance. Avoid non-essential travel to certain areas.",
                    "last_updated": "2025-07-03"
                }
            else:
                return {
                    "success": True,
                    "country": country,
                    "advisory": f"Simulated travel advisory for {country}: Exercise normal precautions. Check local regulations.",
                    "last_updated": "2025-07-03"
                }
        except Exception as e:
            return {"success": False, "error": str(e)}

    def get_chat_completion(self, prompt: str) -> Dict[str, Any]:
        """Gets a chat completion from OpenAI's GPT model."""
        try:
            response = openai.chat.completions.create(
                model="gpt-3.5-turbo",
                messages=[
                    {"role": "system", "content": "You are a helpful assistant."},
                    {"role": "user", "content": prompt}
                ]
            )
            return {"success": True, "response": response.choices[0].message.content}
        except Exception as e:
            return {"success": False, "error": str(e)}


